backend/requirements.txt
Flask
Flask-Cors
Flask-Mail
psycopg2-binary
python-dotenv
gunicorn
bcrypt==4.0.1
sendgrid
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
